requirements.txt
torch
pandas
numpy
transformers
Pillow
matplotlib
scikit-learn
optuna
flask